Refactor helper methods to parent class
